+++ Sharegram/Core/Profile/CurrentUserProfileView.swift
@@ -9,81 +9,21 @@ import SwiftUI
 
 struct CurrentUserProfileView: View {
     
-    private let gridItems: [GridItem] = [
-        .init(.flexible(), spacing: 1),
-        .init(.flexible(), spacing: 1),
-        .init(.flexible(), spacing: 1)
-    ]
+    let user: User
+    
+    var posts: [Post] {
+        return Post.MOCK_POSTS.filter { $0.user?.username == user.username }
+    }
     
     var body: some View {
         NavigationStack {
             ScrollView {
                 // MARK: - Header
-                VStack(spacing: 10) {
-                    // MARK: - User image
-                    HStack {
-                        Image("userImage")
-                            .resizable()
-                            .scaledToFill()
-                            .frame(width: 80, height: 80)
-                            .clipShape(Circle())
-                        
-                        Spacer()
-                        
-                        // MARK: - Stats
-                        HStack(spacing: 8) {
-                            UserStatView(value: 3, title: "Posts")
-                            UserStatView(value: 3, title: "Followers")
-                            UserStatView(value: 3, title: "Following")
-                        }
-                    }
-                    .padding(.horizontal)
-                    .padding(.bottom, 4)
-                    
-                    // MARK: - Username and Name
-                    VStack(alignment: .leading) {
-                        Text("lkshubert")
-                            .font(.footnote)
-                            .fontWeight(.semibold)
-                        
-                        Text("Lucas Hubert")
-                            .font(.footnote)
-                        
-                    }
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(.horizontal)
-                    
-                    // MARK: - Button actions
-                    Button {
-                        
-                    } label: {
-                        Text("Edit Profile")
-                            .font(.subheadline)
-                            .fontWeight(.semibold)
-                            .frame(width: 360, height: 32)
-                            .foregroundColor(.purple)
-                            .overlay(RoundedRectangle(cornerRadius: 6)
-                                .stroke(
-                                    Color.purple,
-                                    lineWidth: 1
-                                )
-                            )
-                    }
-
-                    
-                    Divider()
-                }
-                .padding(.top, 8)
+                ProfileHeaderView(user: user)
                 
                 // MARK: - Post Grid
                 
-                LazyVGrid(columns: gridItems, spacing: 1) {
-                    ForEach(0 ... 15, id: \.self) { index in
-                        Image("userImage")
-                            .resizable()
-                            .scaledToFill()
-                    }
-                }
+                PostGridView(posts: posts)
             }
             .navigationTitle("Profile")
             .navigationBarTitleDisplayMode(.automatic)
@@ -105,6 +45,6 @@ struct CurrentUserProfileView: View {
 
 struct CurrentUserProfileView_Previews: PreviewProvider {
     static var previews: some View {
-        CurrentUserProfileView()
+        CurrentUserProfileView(user: User.MOCK_USERS[0])
     }
 }
